+++ providers/ibm/db2/docs/index.rst
@@ -1,0 +1,94 @@
+ .. Licensed to the Apache Software Foundation (ASF) under one
+    or more contributor license agreements.  See the NOTICE file
+    distributed with this work for additional information
+    regarding copyright ownership.  The ASF licenses this file
+    to you under the Apache License, Version 2.0 (the
+    "License"); you may not use this file except in compliance
+    with the License.  You may obtain a copy of the License at
+
+ ..   http://www.apache.org/licenses/LICENSE-2.0
+
+ .. Unless required by applicable law or agreed to in writing,
+    software distributed under the License is distributed on an
+    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+    KIND, either express or implied.  See the License for the
+    specific language governing permissions and limitations
+    under the License.
+
+``apache-airflow-providers-ibm-db2``
+=====================================
+
+.. toctree::
+    :hidden:
+    :maxdepth: 1
+    :caption: Basics
+
+    Home <self>
+    Changelog <changelog>
+    Security <security>
+
+.. toctree::
+    :hidden:
+    :maxdepth: 1
+    :caption: Guides
+
+    Connection types <connections>
+    Operators <operators>
+
+.. toctree::
+    :hidden:
+    :maxdepth: 1
+    :caption: Resources
+
+    Example DAGs <example-dags>
+    PyPI Repository <https://pypi.org/project/apache-airflow-providers-ibm-db2/>
+    Installing from sources <installing-providers-from-sources>
+
+.. toctree::
+    :hidden:
+    :maxdepth: 1
+    :caption: System tests
+
+    System Tests <_api/tests/system/providers/ibm/db2/index>
+
+.. toctree::
+    :hidden:
+    :maxdepth: 1
+    :caption: Commits
+
+    Detailed list of commits <commits>
+
+
+Package apache-airflow-providers-ibm-db2
+-----------------------------------------
+
+`IBM Db2 <https://www.ibm.com/products/db2>`__
+
+
+Release: 1.0.0
+
+Provider package
+----------------
+
+This package is for the ``ibm.db2`` provider.
+All classes for this package are in the ``airflow.providers.ibm.db2`` python package.
+
+Installation
+------------
+
+You can install this package on top of an existing Airflow 3.0+ installation via
+``pip install apache-airflow-providers-ibm-db2``
+
+Requirements
+------------
+
+The minimum Apache Airflow version supported by this provider package is ``3.0.0``.
+
+=======================================  ==================
+PIP package                              Version required
+=======================================  ==================
+``apache-airflow``                       ``>=3.0.0``
+``apache-airflow-providers-common-sql``  ``>=1.18.0``
+``ibm-db``                               ``>=3.0.0``
+``ibm-db-sa``                            ``>=0.4.0``
+=======================================  ==================
