+++ providers/ibm/db2/docs/security.rst
@@ -1,0 +1,28 @@
+ .. Licensed to the Apache Software Foundation (ASF) under one
+    or more contributor license agreements.  See the NOTICE file
+    distributed with this work for additional information
+    regarding copyright ownership.  The ASF licenses this file
+    to you under the Apache License, Version 2.0 (the
+    "License"); you may not use this file except in compliance
+    with the License.  You may obtain a copy of the License at
+
+ ..   http://www.apache.org/licenses/LICENSE-2.0
+
+ .. Unless required by applicable law or agreed to in writing,
+    software distributed under the License is distributed on an
+    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+    KIND, either express or implied.  See the License for the
+    specific language governing permissions and limitations
+    under the License.
+
+Security
+========
+
+.. note::
+    If you find a security issue, please follow the `Apache Airflow Security Policy <https://airflow.apache.org/docs/apache-airflow/stable/security/index.html>`__.
+
+Releasing security patches
+--------------------------
+
+This provider package follows the same security policy as the main Apache Airflow project.
+For details, see `Releasing security patches <https://github.com/apache/airflow/blob/main/contributing-docs/12_airflow_security.rst>`__.
