+++ providers/ibm/db2/provider.yaml
@@ -16,7 +16,7 @@
 # under the License.
 
 ---
-package-name: apache-airflow-providers-db2
+package-name: apache-airflow-providers-ibm-db2
 name: IBM Db2
 description: |
   `IBM Db2 <https://www.ibm.com/products/db2>`__
@@ -43,13 +43,13 @@ integrations:
 hooks:
   - integration-name: IBM Db2
     python-modules:
-      - airflow.providers.db2.hooks.db2
+      - airflow.providers.ibm.db2.hooks.db2
 
 operators:
   - integration-name: IBM Db2
     python-modules:
-      - airflow.providers.db2.operators.db2
+      - airflow.providers.ibm.db2.operators.db2
 
 connection-types:
-  - hook-class-name: airflow.providers.db2.hooks.db2.Db2Hook
+  - hook-class-name: airflow.providers.ibm.db2.hooks.db2.Db2Hook
     connection-type: Db2
