+++ providers/ibm/db2/pyproject.toml
@@ -20,9 +20,9 @@ requires = ["flit_core==3.12.0"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "apache-airflow-providers-db2"
+name = "apache-airflow-providers-ibm-db2"
 version = "1.0.0"
-description = "Provider package apache-airflow-providers-db2 for Apache Airflow"
+description = "Provider package apache-airflow-providers-ibm-db2 for Apache Airflow"
 readme = "README.rst"
 license = "Apache-2.0"
 license-files = ['LICENSE', 'NOTICE']
@@ -32,7 +32,7 @@ authors = [
 maintainers = [
     {name="Apache Software Foundation", email="dev@airflow.apache.org"},
 ]
-keywords = [ "airflow-provider", "db2", "airflow", "integration" ]
+keywords = [ "airflow-provider", "ibm", "db2", "airflow", "integration" ]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Environment :: Console",
@@ -85,8 +85,8 @@ apache-airflow-providers-common-sql = {workspace = true}
 apache-airflow-providers-standard = {workspace = true}
 
 [project.urls]
-"Documentation" = "https://airflow.apache.org/docs/apache-airflow-providers-db2/1.0.0"
-"Changelog" = "https://airflow.apache.org/docs/apache-airflow-providers-db2/1.0.0/changelog.html"
+"Documentation" = "https://airflow.apache.org/docs/apache-airflow-providers-ibm-db2/1.0.0"
+"Changelog" = "https://airflow.apache.org/docs/apache-airflow-providers-ibm-db2/1.0.0/changelog.html"
 "Bug Tracker" = "https://github.com/apache/airflow/issues"
 "Source Code" = "https://github.com/apache/airflow"
 "Slack Chat" = "https://s.apache.org/airflow-slack"
@@ -94,7 +94,7 @@ apache-airflow-providers-standard = {workspace = true}
 "YouTube" = "https://www.youtube.com/channel/UCSXwxpWZQ7XZ1WL3wqevChA/"
 
 [project.entry-points."apache_airflow_provider"]
-provider_info = "airflow.providers.db2.get_provider_info:get_provider_info"
+provider_info = "airflow.providers.ibm.db2.get_provider_info:get_provider_info"
 
 [tool.flit.module]
-name = "airflow.providers.db2"
+name = "airflow.providers.ibm.db2"
