+++ providers/ibm/db2/src/airflow/providers/ibm/__init__.py
@@ -14,10 +14,6 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Db2 hooks."""
-
 from __future__ import annotations
 
-__all__ = ["Db2Hook"]
-
-from airflow.providers.db2.hooks.db2 import Db2Hook
+__path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
