+++ scripts/ci/docker-compose/remove-sources.yml
@@ -64,7 +64,6 @@ services:
       - ../../../empty:/opt/airflow/providers/common/sql/src
       - ../../../empty:/opt/airflow/providers/databricks/src
       - ../../../empty:/opt/airflow/providers/datadog/src
-      - ../../../empty:/opt/airflow/providers/db2/src
       - ../../../empty:/opt/airflow/providers/dbt/cloud/src
       - ../../../empty:/opt/airflow/providers/dingding/src
       - ../../../empty:/opt/airflow/providers/discord/src
@@ -81,7 +80,11 @@ services:
       - ../../../empty:/opt/airflow/providers/grpc/src
       - ../../../empty:/opt/airflow/providers/hashicorp/src
       - ../../../empty:/opt/airflow/providers/http/src
+<<<<<<< HEAD
       - ../../../empty:/opt/airflow/providers/ibm/mq/src
+=======
+      - ../../../empty:/opt/airflow/providers/ibm/db2/src
+>>>>>>> 5a055cded9 (Rename Db2 provider to IBM Db2 provider)
       - ../../../empty:/opt/airflow/providers/imap/src
       - ../../../empty:/opt/airflow/providers/influxdb/src
       - ../../../empty:/opt/airflow/providers/informatica/src
