+++ scripts/ci/docker-compose/tests-sources.yml
@@ -77,7 +77,6 @@ services:
       - ../../../providers/common/sql/tests:/opt/airflow/providers/common/sql/tests
       - ../../../providers/databricks/tests:/opt/airflow/providers/databricks/tests
       - ../../../providers/datadog/tests:/opt/airflow/providers/datadog/tests
-      - ../../../providers/db2/tests:/opt/airflow/providers/db2/tests
       - ../../../providers/dbt/cloud/tests:/opt/airflow/providers/dbt/cloud/tests
       - ../../../providers/dingding/tests:/opt/airflow/providers/dingding/tests
       - ../../../providers/discord/tests:/opt/airflow/providers/discord/tests
@@ -94,7 +93,11 @@ services:
       - ../../../providers/grpc/tests:/opt/airflow/providers/grpc/tests
       - ../../../providers/hashicorp/tests:/opt/airflow/providers/hashicorp/tests
       - ../../../providers/http/tests:/opt/airflow/providers/http/tests
+<<<<<<< HEAD
       - ../../../providers/ibm/mq/tests:/opt/airflow/providers/ibm/mq/tests
+=======
+      - ../../../providers/ibm/db2/tests:/opt/airflow/providers/ibm/db2/tests
+>>>>>>> 5a055cded9 (Rename Db2 provider to IBM Db2 provider)
       - ../../../providers/imap/tests:/opt/airflow/providers/imap/tests
       - ../../../providers/influxdb/tests:/opt/airflow/providers/influxdb/tests
       - ../../../providers/informatica/tests:/opt/airflow/providers/informatica/tests
